DEBUG

+++ NearVANA.sh
@@ -217,18 +217,25 @@ if [ ! -f ${PID}.Cleaning.ok ]; then
 			touch ${PID}_R0.Unsubstracted.fastq
 			
 			for VARNAME in "${PLATE[@]}"; do
+				echo ${VARNAME}
 				for R in R1 R2 R0; do
+					echo ${R}
 					split -l 100000000 ${PID}_${VARNAME}_${R}.fastq.trim.deinterlaced TOMERGE_${PID}_${VARNAME}_${R}_
-					for PART in TOMERGE_${PID}_${VARNAME}_${R}_* ; do
-						cat ${PART} >> ${PID}_${R}.Unsubstracted.fastq
-					done
-					rm TOMERGE_${PID}_${VARNAME}_${R}_*
+					if [ -f TOMERGE_${PID}_${VARNAME}_${R}_aa ]; then
+						echo "if"
+						for PART in TOMERGE_${PID}_${VARNAME}_${R}_* ; do
+							echo ${PART}
+							cat ${PART} >> ${PID}_${R}.Unsubstracted.fastq
+						done
+						rm TOMERGE_${PID}_${VARNAME}_${R}_*
+					fi
 				done
 				#cat ${PID}_${VARNAME}_R1.fastq.trim.deinterlaced >> ${PID}_R1.Unsubstracted.fastq
 				#cat ${PID}_${VARNAME}_R2.fastq.trim.deinterlaced >> ${PID}_R2.Unsubstracted.fastq
 				#cat ${PID}_${VARNAME}_R0.fastq.trim.deinterlaced >> ${PID}_R0.Unsubstracted.fastq
 				rm ${PID}_${VARNAME}_R*.fastq.trim.deinterlaced
 			done
+			echo "Done"
 			touch ${PID}.Substraction-Deinterlacing.ok
 		else
 			echo -e "\t- ${PID}.Substraction-Deinterlacing.ok existing, pas"
